+++ lqmt/lqm/parsers/FlexTransform/parser.py
@@ -38,9 +38,11 @@ class FlexTransformParser(object):
         :param parserName: Name of the parser being added
         :param parserConfiguration: Path to the configuration file being added
         """
+        incompat_list = ['STIXparserConfig', 'RuleParserConfig']
         if parserName not in self._transform.Parsers:
-            config_file = open(os.path.join(self._current_dir, parserConfiguration), 'r')
-            self._transform.add_parser(parserName, config_file)
+            if parserName not in incompat_list:  # TODO: added to skip different parser config
+                config_file = open(os.path.join(self._current_dir, parserConfiguration), 'r')
+                self._transform.add_parser(parserName, config_file)
 
     def parse(self, datafile, meta=None):
         """